+ Fixed in-game players from not being able to execute commands. + Created CommandException
