--- src/main/java/io/github/znetworkw/znpcservers/commands/CommandExecuteException.java
@@ -1,7 +1,0 @@
-package io.github.znetworkw.znpcservers.commands;
-
-public class CommandExecuteException extends Exception {
-    public CommandExecuteException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
