--- src/main/java/io/github/znetworkw/znpcservers/commands/CommandNotFoundException.java
@@ -1,7 +1,0 @@
-package io.github.znetworkw.znpcservers.commands;
-
-public class CommandNotFoundException extends Exception {
-    public CommandNotFoundException(String message) {
-        super(message);
-    }
-}
