--- src/main/java/io/github/znetworkw/znpcservers/commands/CommandPermissionException.java
@@ -1,7 +1,0 @@
-package io.github.znetworkw.znpcservers.commands;
-
-public class CommandPermissionException extends Exception {
-    public CommandPermissionException(String message) {
-        super(message);
-    }
-}
