+++ src/main/java/io/github/znetworkw/znpcservers/commands/exception/CommandException.java
@@ -1,0 +1,21 @@
+package io.github.znetworkw.znpcservers.commands.exception;
+
+/**
+ * @author xCodiq - 20/04/2023
+ */
+public class CommandException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public CommandException(String message) {
+		super(message);
+	}
+
+	public CommandException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public CommandException(Throwable cause) {
+		super(cause);
+	}
+}
