+++ src/main/java/io/github/znetworkw/znpcservers/commands/exception/CommandExecuteException.java
@@ -1,0 +1,7 @@
+package io.github.znetworkw.znpcservers.commands.exception;
+
+public class CommandExecuteException extends CommandException {
+    public CommandExecuteException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
