+++ src/main/java/io/github/znetworkw/znpcservers/commands/exception/CommandNotFoundException.java
@@ -1,0 +1,7 @@
+package io.github.znetworkw.znpcservers.commands.exception;
+
+public class CommandNotFoundException extends CommandException {
+    public CommandNotFoundException(String message) {
+        super(message);
+    }
+}
