+++ src/main/java/io/github/znetworkw/znpcservers/commands/exception/CommandPermissionException.java
@@ -1,0 +1,7 @@
+package io.github.znetworkw.znpcservers.commands.exception;
+
+public class CommandPermissionException extends CommandException {
+    public CommandPermissionException(String message) {
+        super(message);
+    }
+}
